[SD-194] Move block utility functions from impl to snowdrop-block
